Change to Long

+++ src/intersect.c
@@ -1,12 +1,12 @@
 #include "intersect.h"
 
 
-Paths init_path(int len, int p[], int q[], int r[]){
+Paths init_path(long len, long p[], long q[], long r[]){
     
 
     Paths pt;
-    int arrSize = len*sizeof(int);
-    int arrSizeN = len*sizeof(node);
+    long arrSize = len*sizeof(int);
+    long arrSizeN = len*sizeof(node);
     pt.len = len;
     pt.p = (int*)malloc(arrSize);
     pt.l = (node*)malloc(arrSizeN);
@@ -21,7 +21,7 @@ Paths init_path(int len, int p[], int q[], int r[]){
 
     // Copy l , r (l<=r)
     node L, R;  
-    for(int i=0;i<len;i++){
+    for(long i=0;i<len;i++){
         L.val = q[i]; R.val = r[i];
         L.p = p[i]; R.p = p[i];
         if (L.val>R.val){
@@ -40,7 +40,7 @@ void kill_path(Paths pt){
     free(pt.r);
 }
 
-MemMerge3 allocMemMerge3(int len){
+MemMerge3 allocMemMerge3(long len){
     //Sort Array
     MemMerge3 temp;
     temp.L = (int*)malloc((len+1)*sizeof(int));
@@ -70,7 +70,7 @@ void kill_MemMerge3(MemMerge3 temp){
     free(temp.a3_R);
 }
 
-MemMergeT allocMemMergeT(int len){
+MemMergeT allocMemMergeT(long len){
     MemMergeT temp;
     temp.Ls_L = (node*)malloc((len+1)*sizeof(node));
     temp.Ls_R = (node*)malloc((len+1)*sizeof(node));
@@ -91,22 +91,22 @@ void kill_MemMergeT(MemMergeT temp){
     free(temp.Rs_R);
 }
 
-void MERGE_SORT_FIRSTofThree(Paths pt, int p, int r, MemMerge3 temp){
+void MERGE_SORT_FIRSTofThree(Paths pt, long p, long r, MemMerge3 temp){
     if (p<r){
-        int q = (p+r) / 2;
+        long q = (p+r) / 2;
         MERGE_SORT_FIRSTofThree(pt, p, q, temp);
         MERGE_SORT_FIRSTofThree(pt, q+1, r, temp);
         MERGE_FIRSTofThree(pt, p, q, r, temp);
     }
 }
 
-void MERGE_FIRSTofThree(Paths pt, int p, int q, int r, MemMerge3 temp){
-    int n1 = q - p + 1; // Length of sub-array Left [p,q]
-    int n2 = r - (q+1) + 1; // Length of sub-array right (q,r]
+void MERGE_FIRSTofThree(Paths pt, long p, long q, long r, MemMerge3 temp){
+    long n1 = q - p + 1; // Length of sub-array Left [p,q]
+    long n2 = r - (q+1) + 1; // Length of sub-array right (q,r]
 
     //Sort Array
-    int* L = temp.L;
-    int* R = temp.R;
+    long* L = temp.L;
+    long* R = temp.R;
     //Paired Array 1
     node* a2_L = temp.a2_L;
     node* a2_R = temp.a2_R;
@@ -129,9 +129,9 @@ void MERGE_FIRSTofThree(Paths pt, int p, int q, int r, MemMerge3 temp){
     R[n2] = INT_MAX;
 
     //Dispatch
-    int i = 0;
-    int j = 0;
-    for(int k=p; k<=r;k++){ //from [p,r]
+    long i = 0;
+    long j = 0;
+    for(long k=p; k<=r;k++){ //from [p,r]
         if(L[i] <= R[j]){
             pt.p[k] = L[i];
             pt.l[k] = a2_L[i];
@@ -153,10 +153,10 @@ void sortPaths_P(Paths pt){
     kill_MemMerge3(temp);
 }
 
-int MERGE_SORT_COUNT_INVERSION(Paths pt, int l, int r, MemMergeT temp){
-    int inv=0;
+long MERGE_SORT_COUNT_INVERSION(Paths pt, long l, long r, MemMergeT temp){
+    long inv=0;
     if (l<r){
-        int m = (l+r)/2;
+        long m = (l+r)/2;
         inv += MERGE_SORT_COUNT_INVERSION(pt, l, m, temp);
         inv += MERGE_SORT_COUNT_INVERSION(pt, m+1, r, temp);
         inv += MERGE_COUNT_INVERSION(pt, l, m, r, temp);
@@ -164,11 +164,11 @@ int MERGE_SORT_COUNT_INVERSION(Paths pt, int l, int r, MemMergeT temp){
     return inv;
 }
 
-int MERGE_COUNT_INVERSION(Paths pt, int l, int m, int r, MemMergeT temp){
+long MERGE_COUNT_INVERSION(Paths pt, long l, long m, long r, MemMergeT temp){
 
     //Count inversion
-    int count = m+1;
-    int inv = 0;
+    long count = m+1;
+    long inv = 0;
 
     //Same Item
     node isSame = {
@@ -176,8 +176,8 @@ int MERGE_COUNT_INVERSION(Paths pt, int l, int m, int r, MemMergeT temp){
         .val=0
     };
 
-    int Iden = 0; //Identical Ps on the right
-    int IdenSite = m+1;
+    long Iden = 0; //Identical Ps on the right
+    long IdenSite = m+1;
     if(pt.p[m]==pt.p[m+1]){
         isSame.val=1;//Same item true
         isSame.p = pt.p[m]; // value of identical p
@@ -192,7 +192,7 @@ int MERGE_COUNT_INVERSION(Paths pt, int l, int m, int r, MemMergeT temp){
     }
 
     //Count inversions and Overlapped pairs
-    for (int k=l;k<=m;k++){
+    for (long k=l;k<=m;k++){
         //Find Max{left} > Min{right}
         while(count <= r && pt.r[k].val >= pt.l[count].val){
             //Rule out counted identical P
@@ -214,8 +214,8 @@ int MERGE_COUNT_INVERSION(Paths pt, int l, int m, int r, MemMergeT temp){
 
 
     //Finish sorting
-    int n1 = m - l + 1;
-    int n2 = r - (m+1) + 1;
+    long n1 = m - l + 1;
+    long n2 = r - (m+1) + 1;
 
     // Sort Array
     node* Ls_L = temp.Ls_L;
@@ -237,12 +237,12 @@ int MERGE_COUNT_INVERSION(Paths pt, int l, int m, int r, MemMergeT temp){
     Ls_R[n2].val = INT_MAX;
     Rs_R[n2].val = INT_MAX;
 
-    int Li = 0;
-    int Lj = 0;
-    int Ri = 0;
-    int Rj = 0;
+    long Li = 0;
+    long Lj = 0;
+    long Ri = 0;
+    long Rj = 0;
 
-    for (int k=l;k<=r;k++){
+    for (long k=l;k<=r;k++){
         //Dispatch Ls
         if(Ls_L[Li].val <= Ls_R[Lj].val){
             pt.l[k] = Ls_L[Li];
@@ -268,12 +268,12 @@ int MERGE_COUNT_INVERSION(Paths pt, int l, int m, int r, MemMergeT temp){
 }
 
 
-int get_intersects(int n, int* p, int* q, int* r){
+long get_intersects(long n, long* p, long* q, long* r){
     Paths pt = init_path(n,p,q,r);
     sortPaths_P(pt);
 
     MemMergeT temp = allocMemMergeT(pt.len);
-    int inv = MERGE_SORT_COUNT_INVERSION(pt, 0, pt.len-1, temp);
+    long inv = MERGE_SORT_COUNT_INVERSION(pt, 0, pt.len-1, temp);
     
     kill_MemMergeT(temp);
     kill_path(pt);
@@ -288,10 +288,10 @@ void swapNode(node* a, node* b){
     *b = tmp;
 }
 
-void copyNode(node* dst, node* src, int left, int right){
-    int j = 0;
+void copyNode(node* dst, node* src, long left, long right){
+    long j = 0;
     assert(left<=right);
-    for(int i=left; i<right;i++){
+    for(long i=left; i<right;i++){
         dst[j] = src[i];
         ++j;
     }
